requirements.txt
torch
torchaudio
hydra-core
scikit-learn
pandas
tqdm
pyyaml
soundfile
